+++ src/main/scala/au/com/agl/arc/extract/ORCExtract.scala
@@ -63,6 +63,7 @@ object ORCExtract {
 
     // if incoming dataset has 0 columns then create empty dataset with correct schema
     val emptyDataframeHandlerDF = try {
+      stageDetail.put("records", Integer.valueOf(0))
       ExtractUtils.emptyDataFrameHandler(df, optionSchema)(spark)
     } catch {
       case e: Exception => throw new Exception(e.getMessage) with DetailException {