+++ src/main/scala/au/com/agl/arc/extract/TextExtract.scala
@@ -78,6 +78,7 @@ object TextExtract {
 
     // if incoming dataset has 0 columns then create empty dataset with correct schema
     val emptyDataframeHandlerDF = try {
+      stageDetail.put("records", Integer.valueOf(0))
       ExtractUtils.emptyDataFrameHandler(df, optionSchema)(spark)
     } catch {
       case e: Exception => throw new Exception(e.getMessage) with DetailException {