+++ src/main/scala/au/com/agl/arc/extract/XMLExtract.scala
@@ -85,6 +85,7 @@ object XMLExtract {
 
     // if incoming dataset has 0 columns then create empty dataset with correct schema
     val emptyDataframeHandlerDF = try {
+      stageDetail.put("records", Integer.valueOf(0))
       ExtractUtils.emptyDataFrameHandler(df, optionSchema)(spark)
     } catch {
       case e: Exception => throw new Exception(e.getMessage) with DetailException {