+++ src/test/scala/au/com/agl/arc/plugins/DynamicConfigurationPluginTest.scala
@@ -3,7 +3,7 @@ import java.util
 
 import au.com.agl.arc.util.log.logger.Logger
 
-class ArcConfigPluginTest extends ConfigPlugin {
+class DynamicConfigurationPluginTest extends DynamicConfigurationPlugin {
   override def values()(implicit logger: Logger): util.Map[String, Object] = {
     val values = new java.util.HashMap[String, Object]()
     values.put("arc.foo", "baz")