+++ src/test/scala/au/com/agl/arc/udf/UDFSuite.scala
@@ -80,5 +80,17 @@ class UDFSuite extends FunSuite with BeforeAndAfter {
 
     assert(df.first.getAs[scala.collection.mutable.WrappedArray[Long]](0)(0) == 2147483648L)
     assert(df.schema.fields(0).dataType.toString == "ArrayType(LongType,false)")
-  }     
+  } 
+
+  test("random") {
+    implicit val spark = session
+
+    val df = spark.sql("""
+    SELECT random() AS test
+    """)
+
+    assert(df.first.getDouble(0) > 0.0)
+    assert(df.first.getDouble(0) < 1.0)
+    assert(df.schema.fields(0).dataType.toString == "ArrayType(DoubleType,false)")
+  }       
 }